\name{hydrostats}
\alias{hydrostats}
\alias{hydrostats}
\docType{package}
\title{
calculate hydrologic statistics
}
\description{
A package to calculate a range of hydrologic statistics based on daily time series data and which are widely used in hydrology and ecological applications.
}
\details{
\tabular{ll}{
Package: \tab hydrostats\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-12-20\cr
License: \tab GPL (>= 2) \cr
}

For most functions data must be provided as a dataframe in which the first column is the date in POSIX format (see \code{\link[hydrostats:ts.format]{ts.format}}) and the second column contains the flow data.

}
\author{
Nick Bond

Maintainer:  nrb.ski@gmail.com

}
\references{
%~~ Literature or other references for background information ~~
}
\keyword{ package 
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
}
\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%~~ simple examples of the most important functions ~~
}
