% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erf.R
\name{erf}
\alias{erf}
\title{Error Function}
\usage{
erf(t)
}
\arguments{
\item{t}{argument}
}
\value{
error function
}
\description{
This function returns the error function.  The complementary
error function can be obtained by subtracting the result
from one: (1-erf(t)).  This uses a truncated infinite series.
}
