% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evi.R
\name{evi}
\alias{evi}
\title{Extreme Value type I}
\usage{
evi(x, Fx)
}
\arguments{
\item{x}{data}

\item{Fx}{Fx non-exceedence probability, F(x)=}
}
\value{
inverse EVI CDF
}
\description{
This function returns the inverse of the Extreme
Value type I cummulative distribution F(x).  Do
not normalize these data, EVI uses unique
geometric statistics.
}
