% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyd.mo.R
\name{hyd.mo}
\alias{hyd.mo}
\title{Converts to hydrologic month from date}
\usage{
hyd.mo(dt, h = "N")
}
\arguments{
\item{dt}{date in year, month, day, POSIX, or lubridate format}

\item{h}{"N"orthern or "S"outhern/tropical region}
}
\value{
hydrologic year
}
\description{
Judges and converts to hydrologic calendar from date.
Note that the hydrologic year starts 01 Oct in the
Northern hemisphere and 01 Aug in the Southern and
tropical regions.
}
