% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pevi.R
\name{pevi}
\alias{pevi}
\title{Extreme Value type I}
\usage{
pevi(d, x)
}
\arguments{
\item{d}{data}

\item{x}{x random variable quantity}
}
\value{
non-exceedence probability, F(x)=
}
\description{
This function returns the cummulative
distribution F(x) of the Extreme Value
type I .  Do not normalize these data,
EVI uses unique geometric statistics.
}
