% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnorminv.R
\name{pnorminv}
\alias{pnorminv}
\title{Inverse Normal Cumulative Distribution Function}
\usage{
pnorminv(p, m = 0, s = 1)
}
\arguments{
\item{p}{probability of nonexceedence < 1}

\item{m}{mean of sample}

\item{s}{standard deviation of sample}
}
\value{
value of nonexceedence for that probability
}
\description{
This function returns the inverse cumulative distribution function
value for the normal distribution.  It is calculated based on the
inverse complemenary error function.  The maximum CDF value that
can be inverted is x = 0.999.
}
