% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych.R
\name{psych}
\alias{psych}
\title{Psychrometric constant as a function of temperature}
\usage{
psych(temp, unit = "C")
}
\arguments{
\item{temp}{temperature}

\item{unit}{temperature unit (C)elcius, (K)elvin, (F)ahrenheit, (R)ankine}
}
\value{
psychrometric constant (Pa/K)
}
\description{
This function returns the psychrometric
constant.  The output is in Pascals per
Kelvin.  This is typically represented in
equations as \gamma
}
