% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt3.R
\name{pt3}
\alias{pt3}
\title{Pearson type III distribution}
\usage{
pt3(c, Fx)
}
\arguments{
\item{c}{skewness of sample}

\item{Fx}{non-exceedence probability, F(x)=}
}
\value{
the inverse CDF or K value
}
\description{
This function returns the inverse CDF of the Pearson
type III distribution, that is, the K value.  It is
calculated based on the approximation of:
Kite (1977), Frequency and Risk Analysis in Hydrology,
Water Resouces Publications, Fourt Collins.
as cited in:
Chow, Maidment, and Mays (1988), Applied Hydrology, ch. 12
}
