% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_curve.R
\name{sat_curve}
\alias{sat_curve}
\title{Slope of the saturation vapor pressure curve as a function of temperature}
\usage{
sat_curve(temp, unit = "C")
}
\arguments{
\item{temp}{temperature}

\item{unit}{temperature unit (C)elcius, (K)elvin, (F)ahrenheit, (R)ankine}
}
\value{
slope of the saturation vapor pressure curve (Pa/K)
}
\description{
This function returns the slope of the
saturation vapor pressure of water in the
atmosphere.  The output is in Pascals per
Kelvin.  This is typically represented in
equations as \Delta
}
