% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdev.R
\name{stdev}
\alias{stdev}
\title{Standard Deviation}
\usage{
stdev(x, biascorr = TRUE)
}
\arguments{
\item{x}{data}

\item{biascorr}{use the bias-corrected equation}
}
\value{
s
}
\description{
This function returns the standard deviation of the
data.  This function is writen to remove NAs and
use the bias-corrected statistics with (n-1) in the
demominator.  The population standard deviation can
be found by forcing the denominator to be (n).  Set
the option to FALSE.
}
