% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterrho.R
\name{waterrho}
\alias{waterrho}
\title{Density as a function of temperature}
\usage{
waterrho(temp, unit = "C")
}
\arguments{
\item{temp}{temperature}

\item{unit}{temperature unit (C)elcius, (K)elvin, (F)ahrenheit, (R)ankine}
}
\value{
water density (kg/m^3)
}
\description{
This function is valid for liquid water at
atmospheric pressure; it does not accout
for changes in atmospheric pressure.  Given
the data from USGS.  Input temperature is
provided in Celcius, but can be adjusted to
F, K, or R.  The output is in kg/m^3.
}
