% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wei.R
\name{wei}
\alias{wei}
\title{Weibull Distrubtion}
\usage{
wei(p, m = 0, s = 1)
}
\arguments{
\item{p}{probability of nonexceedence < 1}

\item{m}{mean of sample}

\item{s}{standard deviation of sample}
}
\value{
value of nonexceedence for that probability
}
\description{
This function returns the inverse cumulative distribution function
value for the Weibull Distribution, or the third asymptote of the
extreme value distribution (type III).  It is calculated based on the
inverse complemenary error function.  The maximum CDF value that
can be inverted is x = 0.999.
}
