% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erfinv.R
\name{erfinv}
\alias{erfinv}
\title{Inverse Error Function}
\usage{
erfinv(z)
}
\arguments{
\item{z}{argument}
}
\value{
inverse error function
}
\description{
This function returns the inverse error function.  The
inverse complementary error function can be obtained
with the arguement of (1-z).  This uses a truncated
infinite series.  If more precision is needed, re-write
the series in the for loop to include more higher-order
terms.
}
