% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt3.R
\name{pt3}
\alias{pt3}
\title{Pearson type III distribution}
\usage{
pt3(m, s, c)
}
\arguments{
\item{m}{mean of sample}

\item{s}{standard deviation of sample}

\item{c}{skewness of sample}
}
\value{
the CDF or K value
}
\description{
This function returns the cumulative distribution function value for
the Pearson type III distribution, that is, the K value.  It is
calculated with the gamma function based on the skewness as an input.
}
